backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
bcrypt==4.1.2
pyjwt==2.8.0
python-multipart==0.0.9
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
PyPDF2==3.0.1
python-docx==1.1.0
openai==1.54.0

